requirements.txt
aiomysql==0.0.9
appdirs==1.4.3
mysql-connector==2.1.4
packaging==16.8
psycopg2==2.7.1
PyMySQL==0.7.11
pyparsing==2.2.0
six==1.10.0
